requirements.txt
setuptools==65.4.1
tensorflow-macos
tensorflow-metal
matplotlib~=3.7.1
pip~=22.2.2
wheel~=0.37.1
Pillow~=9.2.0
packaging~=21.3
pyparsing~=3.0.9
zipp~=3.8.1
numpy~=1.23.3
keras
imageio~=2.26.1